0.7 Finished Home Page
finally finished the home page... added the jumbo, medium, premium, and staff picks ducks. made it so that the navbar will actually navigate to each of the sections with anchor tags. added a footer with a fake address (the ducks aren't exactly how i wanted them, you can mess w them if you want but i got mad trying to do it so), phone number, email, and the little ducks in the corner. added the buy now buttons and made them link to a future made buy now form (will be done later today technically, but i need sleep lol).